refactor(reader/): ConsoleReader를 util/에서 view/reader/로 이동
